Dispatch from within the convolve function
