+++ dask_image/utils/_dispatcher.py
@@ -10,3 +10,21 @@ class Dispatcher(Dispatch):
         """
         meth = self.dispatch(type(arg._meta))
         return meth(arg, *args, **kwargs)
+
+
+convolve_dispatch = Dispatcher(name="convolve_dispatch")
+
+
+@convolve_dispatch.register(np.ndarray)
+def numpy_convolve(*args, **kwargs):
+    return scipy.ndimage.filters.convolve
+
+
+@convolve_dispatch.register_lazy("cupy")
+def register_cupy():
+    import cupy
+    import cupyx.scipy.ndimage
+
+    @convolve_dispatch.register(cupy.ndarray)
+    def cupy_convolve(*args, **kwargs):
+        return cupyx.scipy.ndimage.filters.convolve
